feat/update several height changes to 1024px and move sidebar component to sidebar folder
