+++ src/components/Sidebar/SidebarAdmin.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function SidebarAdmin() {
+  return <div>SidebarAdmin</div>;
+}
