+++ src/pages/Customer/DashboardUser.js
@@ -3,13 +3,13 @@ import BillingDate from "../../components/BillingDate";
 import DoughnutChartSection from "../../components/DoughnutChartSection";
 import History from "../../components/History";
 import Header from "../../components/Header";
-import Sidebar from "../../components/Sidebar";
+import Sidebar from "../../components/Sidebar/Sidebar";
 
 export default function DashboardUser() {
   return (
-    <div className="flex w-screen h-screen App bg-[#F3F3F3]">
+    <div className="flex w-screen h-[1024px] App bg-[#F3F3F3]">
       <Sidebar />
-      <div className="w-[1150px]">
+      <div className="">
         <Header />
         <div className="flex ml-[15px]">
           <DoughnutChartSection />
